+++ mycodo/mycodo_flask/api/input.py
@@ -10,12 +10,15 @@ from flask_restx import fields
 
 from mycodo.databases.models import DeviceMeasurements
 from mycodo.databases.models import Input
+from mycodo.databases.models import InputChannel
+from mycodo.databases.models.input import InputChannelSchema
 from mycodo.databases.models.input import InputSchema
 from mycodo.databases.models.measurement import DeviceMeasurementsSchema
 from mycodo.mycodo_client import DaemonControl
 from mycodo.mycodo_flask.api import api
 from mycodo.mycodo_flask.api import default_responses
 from mycodo.mycodo_flask.api.sql_schema_fields import device_measurement_fields
+from mycodo.mycodo_flask.api.sql_schema_fields import input_channel_fields
 from mycodo.mycodo_flask.api.sql_schema_fields import input_fields
 from mycodo.mycodo_flask.api.utils import get_from_db
 from mycodo.mycodo_flask.api.utils import return_list_of_dictionaries
@@ -27,12 +30,14 @@ ns_input = api.namespace('inputs', description='Input operations')
 
 input_single_fields = api.model('Input Status Fields', {
     'input settings': fields.Nested(input_fields),
+    'input channels': fields.List(fields.Nested(input_channel_fields)),
     'device measurements': fields.List(
         fields.Nested(device_measurement_fields)),
 })
 
 input_list_fields = api.model('Input Fields List', {
     'input settings': fields.List(fields.Nested(input_fields)),
+    'input channels': fields.List(fields.Nested(input_channel_fields))
 })
 
 
@@ -50,8 +55,10 @@ class Inputs(Resource):
             abort(403)
         try:
             list_data = get_from_db(InputSchema, Input)
+            list_channels = get_from_db(InputChannelSchema, InputChannel)
             if list_data:
-                return {'input settings': list_data}, 200
+                return {'input settings': list_data,
+                        'input channels': list_channels}, 200
         except Exception:
             abort(500,
                   message='An exception occurred',
@@ -75,16 +82,23 @@ class SettingsInputsUniqueID(Resource):
         if not utils_general.user_has_permission('view_settings'):
             abort(403)
         try:
-            dict_data = get_from_db(InputSchema, Input, unique_id=unique_id)
+            list_data = get_from_db(InputSchema, Input, unique_id=unique_id)
+
+            measure_schema = InputChannelSchema()
+            list_channels = return_list_of_dictionaries(
+                measure_schema.dump(
+                    InputChannel.query.filter_by(
+                        input_id=unique_id).all(), many=True))
 
             measure_schema = DeviceMeasurementsSchema()
-            list_data = return_list_of_dictionaries(
+            list_measurements = return_list_of_dictionaries(
                 measure_schema.dump(
                     DeviceMeasurements.query.filter_by(
                         device_id=unique_id).all(), many=True))
 
-            return {'input settings': dict_data,
-                    'device measurements': list_data}, 200
+            return {'input settings': list_data,
+                    'input channels': list_channels,
+                    'device measurements': list_measurements}, 200
         except Exception:
             abort(500,
                   message='An exception occurred',